set up google auth2 passport js

+++ controllers/api/v1/posts_api.js
@@ -22,20 +22,21 @@ module.exports.index = async function(req,res){
 
 module.exports.destroy = async function(req,res){
     try {
+
         let post = await Post.findById(req.params.id);
-        //  if(post.user == req.user.id){
+         if(post.user == req.user.id){
             post.remove();
             await Comment.deleteMany({post: req.params.id});
 
         //to pass variables to js file 
-         
            return res.json(200,{
                message: 'post and associated comment deleted'
            })
-        // }else{
-        //     req.flash('error','you cannot delete this post');
-        //         return res.redirect('back');
-        //     }
+        }else{
+            return res.json(401,{
+                message: "you cannot delete this post!"
+            });
+            }
     } catch (err) {
         return res.json(500,{
             message: 'Internal Server Error'
